+++ resources/views/pages/post/create.blade.php
@@ -21,7 +21,7 @@
                             </select>
                             <label for="titulo" class="fs-5">Título</label>
                             <input type="text" id="titulo" name="titulo" class="form-control mb-2" required>
-                            <textarea id="" class="summernote" name="content"></textarea>
+                            <textarea id="" class="summernote summernote_create" name="content"></textarea>
                             <input type="hidden" name="user_id" value="{{ auth()->user()->id }}">
                             <div class="text-center">
                                 <button type="submit" class="btn btn-primary mt-2">Publicar</button>
@@ -38,21 +38,34 @@
     <script>
         document.getElementById('postForm').addEventListener('submit', function(event) {
             event.preventDefault();
-            Swal.showLoading()
-            let formData = new FormData(this);
-            //console.log(formData)
-            axios.post('{{ route('post.store') }}', formData)
-                .then(function(response) {
-                    //console.log(response.data.postUrl)
-                    let postUrl = response.data.postUrl;
-                    let mensaje = response.data.mensaje;
-                    localStorage.setItem('mensaje', mensaje);
-                    window.location.replace(postUrl);
-                })
-                .catch(function(error) {
-                    console.error(error);
-                    console.error(response.data.postUrl);
+            if ($('.summernote_create').summernote('isEmpty')) {
+                //console.log('hola')
+                Swal.fire({
+                    icon: 'error',
+                    title: 'Oops!',
+                    text: '¡Escribe algo!',
                 });
+            }else{
+                Swal.showLoading()
+                let formData = new FormData(this);
+                //console.log(formData)
+                axios.post('{{ route('post.store') }}', formData)
+                    .then(function(response) {
+                        //console.log(response.data.postUrl)
+                        let postUrl = response.data.postUrl;
+                        let mensaje = response.data.mensaje;
+                        localStorage.setItem('mensaje', mensaje);
+                        window.location.replace(postUrl);
+                    })
+                    .catch(function(error) {
+                        console.error(error);
+                        Swal.fire({
+                            icon: 'error',
+                            title: 'Oops!',
+                            text: error.response.data.message,
+                        });
+                    });
+            }
         });
     </script>
 @endpush
